refactor: change folder/files structure
