+++ packages/html-to-slate-ast/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig(options => ({
-  entry: ['src/index.ts', 'src/client'],
+  entry: ['src/index.ts', 'src/client.ts'],
   tsconfig: 'tsconfig.build.json',
   minify: !options.watch,
   sourcemap: true,
